chore: rename project to react-native-bottom-tabs
