+++ android/gradle.properties
@@ -1,5 +1,5 @@
-SwiftuiTabview_kotlinVersion=1.7.0
-SwiftuiTabview_minSdkVersion=21
-SwiftuiTabview_targetSdkVersion=31
-SwiftuiTabview_compileSdkVersion=31
-SwiftuiTabview_ndkversion=21.4.7075529
+RCTTabView_kotlinVersion=1.7.0
+RCTTabView_minSdkVersion=21
+RCTTabView_targetSdkVersion=31
+RCTTabView_compileSdkVersion=31
+RCTTabView_ndkversion=21.4.7075529
