+++ android/src/main/java/com/rcttabview/RCTTabView.kt
@@ -1,4 +1,4 @@
-package com.swiftuitabview
+package com.rcttabview
 
 import android.content.Context
 import android.view.MenuItem
