+++ android/src/main/java/com/rcttabview/RCTTabViewPackage.kt
@@ -1,4 +1,4 @@
-package com.swiftuitabview
+package com.rcttabview
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
@@ -6,10 +6,10 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 import java.util.ArrayList
 
-class SwiftuiTabviewViewPackage : ReactPackage {
+class RCTTabViewPackage : ReactPackage {
   override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
     val viewManagers: MutableList<ViewManager<*, *>> = ArrayList()
-    viewManagers.add(SwiftuiTabviewViewManager())
+    viewManagers.add(RCTTabViewViewManager())
     return viewManagers
   }
