+++ android/src/oldarch/RCTTabViewViewManager.kt
@@ -1,9 +1,9 @@
-package com.swiftuitabview
+package com.rcttabview
 
 import android.view.View
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.SimpleViewManager
 
-abstract class SwiftuiTabviewViewManagerSpec<T : View> : SimpleViewManager<T>() {
+abstract class RCTTabViewViewManagerSpec<T : View> : SimpleViewManager<T>() {
   abstract fun setColor(view: T?, value: String?)
 }
