+++ example/ios/.xcode.env
@@ -1,1 +1,1 @@
-export NODE_BINARY='/Users/okwasniewski/.nvm/versions/node/v20.15.1/bin/node'
+export NODE_BINARY='/opt/homebrew/bin/node'
