+++ example/ios/Podfile
@@ -4,6 +4,6 @@ ws_dir = ws_dir.parent until
   ws_dir.expand_path.to_s == '/'
 require "#{ws_dir}/node_modules/react-native-test-app/test_app.rb"
 
-workspace 'SwiftuiTabviewExample.xcworkspace'
+workspace 'ReactNativeBottomTabsExample.xcworkspace'
 
 use_test_app!
