+++ example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-native-swiftui-tabview-example",
+  "name": "react-native-bottom-tabs-example",
   "version": "0.0.1",
   "private": true,
   "scripts": {
