+++ ios/Fabric/RCTTabViewView.h
@@ -3,15 +3,15 @@
 #import <React/RCTViewComponentView.h>
 #import <UIKit/UIKit.h>
 
-#ifndef SwiftuiTabviewViewNativeComponent_h
-#define SwiftuiTabviewViewNativeComponent_h
+#ifndef RCTTabViewView_h
+#define RCTTabViewView_h
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface SwiftuiTabviewView : RCTViewComponentView
+@interface RCTTabViewView : RCTViewComponentView
 @end
 
 NS_ASSUME_NONNULL_END
 
-#endif /* SwiftuiTabviewViewNativeComponent_h */
+#endif /* RCTTabViewView_h */
 #endif /* RCT_NEW_ARCH_ENABLED */
